Sync fork to upstream parent

+++ src/Certify.UI/ViewModel/AppViewModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.Linq;
+using System.Reflection;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Windows.Input;
@@ -214,10 +215,6 @@ namespace Certify.UI.ViewModel
 
         public async Task InitServiceConnections()
         {
-            // wire up stream events
-            CertifyClient.OnMessageFromService += CertifyClient_SendMessage;
-            CertifyClient.OnRequestProgressStateUpdated += UpdateRequestTrackingProgress;
-            CertifyClient.OnManagedCertificateUpdated += CertifyClient_OnManagedCertificateUpdated;
 
             //check service connection
             IsServiceAvailable = await CheckServiceAvailable();
@@ -226,15 +223,26 @@ namespace Certify.UI.ViewModel
             {
                 Debug.WriteLine("Service not yet available. Waiting a few seconds..");
 
-                // the service could still be starting up
+                // the service could still be starting up or port may be reallocated
                 await Task.Delay(5000);
+
+
+                // restart client in case port has reallocated
+                CertifyClient = new CertifyServiceClient();
+
                 IsServiceAvailable = await CheckServiceAvailable();
+
                 if (!IsServiceAvailable)
                 {
                     // give up
                     return;
                 }
             }
+
+            // wire up stream events
+            CertifyClient.OnMessageFromService += CertifyClient_SendMessage;
+            CertifyClient.OnRequestProgressStateUpdated += UpdateRequestTrackingProgress;
+            CertifyClient.OnManagedCertificateUpdated += CertifyClient_OnManagedCertificateUpdated;
 
             // connect to status api stream & handle events
             await CertifyClient.ConnectStatusStreamAsync();
@@ -249,11 +257,17 @@ namespace Certify.UI.ViewModel
               });
         }
 
+        /// <summary>
+        /// Checks the service availability by fetching the version. If the service is available but the version is wrong an exception will be raised.
+        /// </summary>
+        /// <returns></returns>
         public async Task<bool> CheckServiceAvailable()
         {
+            string version = null;
             try
             {
-                await CertifyClient.GetAppVersion();
+                version = await CertifyClient.GetAppVersion();
+
                 IsServiceAvailable = true;
             }
             catch (Exception)
@@ -262,7 +276,27 @@ namespace Certify.UI.ViewModel
                 IsServiceAvailable = false;
             }
 
-            return IsServiceAvailable;
+            if (version != null)
+            {
+
+                // ensure service is correct version
+                var v = Version.Parse(version.Replace("\"", ""));
+
+                var assemblyVersion = typeof(AppViewModel).Assembly.GetName().Version;
+
+                if (v.Major != assemblyVersion.Major)
+                {
+                    throw new Exception($"Invalid service version ({v}). Please ensure the old version of the app has been fully uninstalled, then re-install the latest version.");
+                }
+                else
+                {
+                    return IsServiceAvailable;
+                }
+            }
+            else
+            {
+                return IsServiceAvailable;
+            }
         }
 
         /// <summary>
@@ -321,7 +355,7 @@ namespace Certify.UI.ViewModel
             return false;
         }
 
-        public async Task BeginCertificateRequest(string managedItemId)
+        public async Task BeginCertificateRequest(string managedItemId, bool resumePaused = true)
         {
             //begin request process
             var managedCertificate = ManagedCertificates.FirstOrDefault(s => s.Id == managedItemId);
@@ -332,7 +366,7 @@ namespace Certify.UI.ViewModel
 
                 TrackProgress(managedCertificate);
                 // start request
-                var result = await CertifyClient.BeginCertificateRequest(managedCertificate.Id);
+                var result = await CertifyClient.BeginCertificateRequest(managedCertificate.Id, resumePaused);
             }
         }
 
@@ -446,6 +480,8 @@ namespace Certify.UI.ViewModel
 
         public async Task<bool> DeleteCredential(string credentialKey)
         {
+            if (credentialKey == null) return false;
+
             var result = await CertifyClient.DeleteCredential(credentialKey);
             await RefreshStoredCredentialsList();
